+++ src/debops/projectdir.py
@@ -13,6 +13,9 @@ import socket
 import distro
 import platform
 import pathlib
+import git
+import subprocess
+import time
 
 
 class ProjectDir(object):
@@ -93,6 +96,11 @@ class ProjectDir(object):
 
         self.config.merge(os.path.join(self.path, '.debops', 'conf.d'))
 
+        self._commands = {
+            'ansible-galaxy': self.config.raw['binaries']['ansible-galaxy'],
+            'git-crypt': self.config.raw['binaries']['git-crypt']
+        }
+
         # Set the default view
         try:
             self.view = self.config.raw['project']['default_view']
@@ -150,12 +158,25 @@ class ProjectDir(object):
             self.ansible_cfg = AnsibleConfig(
                     os.path.join(self.path, 'ansible.cfg'),
                     project_type=self.project_type)
+            if not self.config.get_env('DEBOPS_ANSIBLE_INVENTORY'):
+                self.config.set_env('DEBOPS_ANSIBLE_INVENTORY',
+                                    unexpanduser(os.path.join(self.path,
+                                                              'ansible',
+                                                              'inventory')))
+
         elif self.project_type == 'modern':
             self.ansible_cfg = AnsibleConfig(
                     os.path.join(self.path, 'ansible', 'views',
                                  self.view, 'ansible.cfg'),
                     project_type=self.project_type,
                     view=self.view)
+            if not self.config.get_env('DEBOPS_ANSIBLE_INVENTORY'):
+                self.config.set_env('DEBOPS_ANSIBLE_INVENTORY',
+                                    unexpanduser(os.path.join(self.path,
+                                                              'ansible',
+                                                              'views', self.view,
+                                                              'inventory')))
+
         self.ansible_cfg.load_config()
         self.config.set_env('ANSIBLE_CONFIG',
                             unexpanduser(self.ansible_cfg.path))
@@ -188,6 +209,14 @@ class ProjectDir(object):
                 return path
             path = os.path.dirname(last_path)
         return None
+
+    def _is_git_repo(self, path):
+        try:
+            _ = git.Repo(path).git_dir
+            return True
+        except (git.exc.InvalidGitRepositoryError,
+                git.exc.NoSuchPathError):
+            return False
 
     def _write_file(self, filename, *content):
         """
@@ -505,6 +534,52 @@ class ProjectDir(object):
         elif self.project_type == 'legacy':
             self._create_legacy_project(self.path)
 
+        create_git_repo = self.kwargs.get('git', None)
+        if create_git_repo:
+            repo = git.Repo.init(self.path)
+            repo.git.add(all=True)
+            repo.index.commit(self.config.raw['git']['init_message'])
+
+            # Set up encryption using git-crypt
+            encrypt_git_repo = self.kwargs.get('encrypt', None)
+            if encrypt_git_repo == 'git-crypt':
+                try:
+                    gpg_keys = list(self.kwargs.get('keys', None).split(','))
+                except AttributeError:
+                    raise ValueError('List of GPG recipients not specified')
+
+                os.chdir(self.path)
+                gitcrypt_cmd = subprocess.Popen([self._commands['git-crypt'],
+                                                 'init'],
+                                                stdin=subprocess.PIPE)
+                gitcrypt_cmd.communicate()
+                while not os.path.exists(
+                        os.path.join('.git', 'git-crypt', 'keys')):
+                    time.sleep(1)
+                for gpg_key in gpg_keys:
+                    gitcrypt_cmd = subprocess.Popen([self._commands['git-crypt'],
+                                                     'add-gpg-user',
+                                                    gpg_key], stdin=subprocess.PIPE)
+                    gitcrypt_cmd.communicate()
+
+                # Lock the repository after setting up git-crypt
+                gitcrypt_cmd = subprocess.Popen([self._commands['git-crypt'],
+                                                 'lock'],
+                                                stdin=subprocess.PIPE)
+                gitcrypt_cmd.communicate()
+
+            # Install Ansible Collections after the project is initialized
+            install_requirements = self.kwargs.get('requirements', None)
+            if install_requirements:
+                os.chdir(self.path)
+                galaxy_cmd = subprocess.Popen([self._commands['ansible-galaxy'],
+                                               'collection', 'install', '-r',
+                                               os.path.join('ansible',
+                                                            'collections',
+                                                            'requirements.yml')],
+                                              stdin=subprocess.PIPE)
+                galaxy_cmd.communicate()
+
     def mkview(self, view):
 
         # Make sure that users are not trying to nest the view inside of
@@ -625,6 +700,28 @@ class ProjectDir(object):
         else:
             raise NotADirectoryError('This functionality only works in '
                                      '"modern" DebOps project directory')
+
+    def commit(self, interactive=False):
+        """Commit the current contents of the project directory to the git
+        repository automatically."""
+        if self._is_git_repo(self.path):
+            repo = git.Repo(self.path)
+            repo.git.add(all=True)
+
+            # Check if there are any differences between the current HEAD and
+            # the index. If there are, we need to commit them.
+            diff_list = repo.head.commit.diff()
+            if diff_list:
+                try:
+                    repo.index.commit(
+                        self.config.raw['project']['git']['auto_commit_message'])
+                except KeyError:
+                    # There was an issue in the configuration, unstage any
+                    # changes in git index
+                    repo.git.reset()
+                    if interactive:
+                        print('The "project.git.auto_commit_message" option is '
+                              'not defined. Not committing any changes.')
 
     def refresh(self):
         if self.project_type == 'modern':